backend/requirements.txt
fastapi
uvicorn
python-dotenv
speechrecognition
pyttsx3
pillow
langchain
google-generativeai
langchain-google-genai
opencv-python
rospy
cv_bridge
pydub 